Create Abstract result for item and tes tresult
